requirements.txt
python-telegram-bot==13.7
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
